requirements.txt
django
biopython

#nice forms
django-crispy-forms

# running calculations
rq
django-rq

# user authentication
pyjwkest
bcrypt
django-allauth


# data visualization
numpy
bokeh
